docker-compose: Disable sensor container
We originally used a set of individual I2C sensors but this wasn't
terribly reliable. Now we've moved over to an integrated sensor
board which is the Pimoroni Enviro+. As such I'm disabling this
sensor container.

We use the main Raspberry Pi 4 board for A/V in the container
here. Then we use an attached Raspberry Pi Zero board to run
the sensor code. The idea is that the Pi0 can be run alone
without the more heavyweight Pi4.

For the sensor code see:

https://github.com/DynamicDevices/enviroplus-docker-image

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>